patches/next: Remove 'Subject:' lines from cover-letter patches
'git am' applies 'Subject:' line on bodies of patches if exist.
As a result, cover-letter patches having the lines on bodies changes the
title of the commit when applied via 'git am' or 'assemble_tree.sh'.
Fix the problem by removing the 'Subject:' lines on the commit messages
of patches.

Signed-off-by: SeongJae Park <sj@kernel.org>