scripts/rebase_damon_next_on_mm_unstable: do not copy unmerged_commits.sh on working dir
rebase_damon_next_on_mm_unstable.sh copies unmerged_commits.sh on the working
directory first, and then uses it.  unmerged_commits.sh is not dependent to the
place of the binary, though, so no need to be copied.  The copy only makes the
working directory dirty.  Remove the unnecessary copying.

Signed-off-by: SeongJae Park <sj@kernel.org>