patches/next: remove reverts of reverts of DAMON debugfs interface
Until DAMON debugfs interface removal is merged into the mainline, we
want to keep testing and fixing bugs in DAMON debugfs interface, though
it is deprecated.  For that, we added reverts of the removal on
damon/next tree.  Because we should also test the removal changes, I
added reverts of the reverts for only testing purpose, and mistakenly
added those to the patches queue.  Remove the mitakenly added patches.

Signed-off-by: SeongJae Park <sj@kernel.org>