patches/next: samples/damon/prcl: fix type of max_age_region
Assembled tree: 74b866938a402fc479938747f93b9ced3c9a1282

The type of ->max_age_region of struct damos_access_pattern is unsigned
int, but the sample program is setting it as the max value of unsigned
long type.  Fix it to use the max value of unsigned int type.

Fixes: 18ab72bebbc8 ("patches/next: cleanup sample modules")
Reported-by: kernel test robot <lkp@intel.com>
Closes: https://lore.kernel.org/oe-kbuild-all/202411231420.di7hRYuZ-lkp@intel.com/
Signed-off-by: SeongJae Park <sj@kernel.org>

Signed-off-by: SeongJae Park <sj@kernel.org>