patches/next: drop debug sysfs file
Assembled tree: 69d225b1639d462cf7e8419ef0088fe2d4d92c8f

The patch was added for debugging-purpose show of memcg path to memcg id
mappings.  It is not really being used now.  And it recently only caused
a noise [1].  I promised to fix it on damon/next, but maybe this is
better to just be dropped for now, to avoid similar noises in future.
Drop it.

[1] https://lore.kernel.org/20251227221206.282703-1-sj@kernel.org

Signed-off-by: SeongJae Park <sj@kernel.org>